# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Lotmark",
    "version": "1",
    "description": "ISO 17034 reference material producer platform.\n\nErrors are RFC 9457 problem details. The `code` field is the machine-readable one; `title` is that code humanised for display and must not be branched on. Where a refusal was recorded, `auditSeq` points at the ledger entry.\n\nThis document is generated from a registry that a test compares against the routes the server actually registered, in both directions. It cannot describe an endpoint that does not exist, and an endpoint cannot exist without appearing here."
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local development"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "lotmark_session",
        "description": "An opaque server-side session. Cookies rather than a bearer token because this domain needs IMMEDIATE revocation — a locked account, an idle timeout, a withdrawn competence — and a stateless token cannot be withdrawn before it expires."
      }
    },
...
  },
  "paths": {
...
    "/api/v1/capa": {
      "get": {
        "operationId": "getCapa",
        "summary": "Complaints and CAPAs",
        "description": "Requires `capa:manage`.",
        "tags": [
          "Quality"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "The request succeeded."
          },
          "400": {
            "description": "The request was malformed. RFC 9457 problem details."
          },
          "401": {
            "description": "Not authenticated, the second factor is outstanding, or a signing step-up is required. The `code` distinguishes them."
          },
          "403": {
            "description": "Refused. `code` names the reason — permission_denied, segregation_of_duties, competence_missing or competence_expired — and `auditSeq` points at the ledger entry that recorded the refusal."
          }
        }
      }
    },
    "/api/v1/capa/workflow": {
      "get": {
        "operationId": "getCapaWorkflow",
        "summary": "The declared CAPA state machine",
        "description": "The console renders transitions from this rather than deciding for itself. Requires `capa:manage`.",
        "tags": [
          "Quality"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "The request succeeded."
          },
          "400": {
            "description": "The request was malformed. RFC 9457 problem details."
          },
          "401": {
            "description": "Not authenticated, the second factor is outstanding, or a signing step-up is required. The `code` distinguishes them."
          },
          "403": {
            "description": "Refused. `code` names the reason — permission_denied, segregation_of_duties, competence_missing or competence_expired — and `auditSeq` points at the ledger entry that recorded the refusal."
          }
        }
      }
    },
    "/api/v1/capa/{id}/transition": {
      "post": {
        "operationId": "postCapaByIdTransition",
        "summary": "Move a CAPA to its next state",
        "description": "Requires `capa:manage`.",
        "tags": [
          "Quality"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The id."
          }
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "The request succeeded."
          },
          "400": {
            "description": "The request was malformed. RFC 9457 problem details."
          },
          "401": {
            "description": "Not authenticated, the second factor is outstanding, or a signing step-up is required. The `code` distinguishes them."
          },
          "403": {
            "description": "Refused. `code` names the reason — permission_denied, segregation_of_duties, competence_missing or competence_expired — and `auditSeq` points at the ledger entry that recorded the refusal."
          }
        }
      }
    },
...
  }
}